+++ src/node/ext/ldap/tests/test_ugm_group_of_names.py
@@ -56,6 +56,13 @@ class TestUGMGroupOfNames(NodeTestCase):
         )
         self.assertEqual(str(err), "'inexistent'")
 
+        # Invalidate
+        self.assertEqual(ugm.storage.keys(), ['users', 'groups'])
+        ugm.invalidate(key='users')
+        self.assertEqual(ugm.storage.keys(), ['groups'])
+        ugm.invalidate()
+        self.assertEqual(ugm.storage.keys(), [])
+
     @group_of_names_ugm
     def test_fetch_users(self, ugm):
         # User keys
@@ -89,19 +96,19 @@ class TestUGMGroupOfNames(NodeTestCase):
             'foo',
             object()
         )
-        self.assertEqual(str(err), 'User does not implement ``__setitem__``')
+        self.assertEqual(str(err), 'User does not support ``__setitem__``')
         err = self.expect_error(
             NotImplementedError,
             user_0.__delitem__,
             'foo'
         )
-        self.assertEqual(str(err), 'User does not implement ``__delitem__``')
+        self.assertEqual(str(err), 'User does not support ``__delitem__``')
         err = self.expect_error(
             NotImplementedError,
             user_0.__getitem__,
             'foo'
         )
-        self.assertEqual(str(err), 'User does not implement ``__getitem__``')
+        self.assertEqual(str(err), 'User does not support ``__getitem__``')
         self.assertEqual(user_0.keys(), [])
 
     @group_of_names_ugm
@@ -268,7 +275,7 @@ class TestUGMGroupOfNames(NodeTestCase):
             'uid0',
             users['uid0']
         )
-        self.assertEqual(str(err), 'Group does not implement ``__setitem__``')
+        self.assertEqual(str(err), 'Group does not support ``__setitem__``')
 
         # Members are added via ``add``
         group_1.add('uid0')