+++ src/node/ext/ldap/ugm/_api.py
@@ -1215,4 +1215,9 @@ class LDAPUgm(UgmBase):
     OdictStorage,
 )
 class Ugm(object):
-    pass
+
+    def invalidate(self, key=None):
+        if key is None:
+            self.storage.clear()
+            return
+        del self.storage[key]